src/requirements.txt
streamlit
requests
python-dotenv
spacy
sentence-transformers
faiss-cpu
umap-learn
plotly
